Code cleaning and mild restructuring
