+++ oss-cloud-backend/serverless.yml
@@ -39,12 +39,6 @@ custom:
       convertEmptyValues: true
 
 functions:
-  hello:
-    handler: handler.hello
-    events:
-      - http:
-          path: hello
-          method: get
   addContributor:
     handler: handler.addContributor
     events:
