+++ oss-cloud-frontend/src/axiosService.js
@@ -2,6 +2,7 @@ import axios from "axios";
 
 const addContributorURL = "http://localhost:3000/addContributor";
 const getAllContributorsURL = "http://localhost:3000/contributors";
+const contributionsURL = "http://localhost:3000/contributions";
 
 export const addContributorAxios = contributor => {
   return axios({
@@ -17,3 +18,10 @@ export const loadContributorsAxios = () => {
     url: getAllContributorsURL
   });
 };
+
+export const loadContributionsAxios = () => {
+  return axios({
+    method: 'get',
+    url: contributionsURL,
+  })
+}
